+++ internal/cli/performanceadvisor/namespaces/description.go
@@ -11,23 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+package namespaces
 
-package performanceadvisor
-
-import (
-	"github.com/mongodb/mongocli/internal/cli/atlas/performanceadvisor/namespaces"
-	"github.com/spf13/cobra"
+const (
+	short = "Retrieve namespaces for collections experiencing slow queries"
+	list  = "List namespaces for collections experiencing slow queries"
 )
-
-func Builder() *cobra.Command {
-	const use = "performanceAdvisor"
-	cmd := &cobra.Command{
-		Use:   use,
-		Short: short,
-	}
-	cmd.AddCommand(
-		namespaces.Builder(),
-	)
-
-	return cmd
-}
