+++ internal/cli/performanceadvisor/namespaces/list_test.go
