+++ internal/cli/performanceadvisor/namespaces/namespaces.go
