+++ internal/cli/performanceadvisor/performanceadvisor.go
@@ -11,9 +11,25 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package namespaces
 
-const (
-	short = "Namespaces operations"
-	list  = "List Namespaces"
+package performanceadvisor
+
+import (
+	"github.com/mongodb/mongocli/internal/cli"
+	"github.com/mongodb/mongocli/internal/cli/performanceadvisor/namespaces"
+	"github.com/spf13/cobra"
 )
+
+func Builder() *cobra.Command {
+	const use = "performanceAdvisor"
+	cmd := &cobra.Command{
+		Use:     use,
+		Aliases: cli.GenerateAliases(use),
+		Short:   short,
+	}
+	cmd.AddCommand(
+		namespaces.Builder(),
+	)
+
+	return cmd
+}
